MGDOBR-1275: use a simple flow in ManagedProcessor
